# .well-known/openapi.yaml
openapi: 3.0.1
info:
  title: Google Services Plugin
  description: A plugin that allows ChatGPT to interact with Google Calendar, Tasks, Gmail, and Drive
  version: 'v1'
servers:
  - url: https://ucandoai.netlify.app/.netlify/functions/api
paths:
  /api/calendar/events:
    get:
      operationId: getCalendarEvents
      summary: Get upcoming calendar events
      description: Retrieves a list of upcoming events from the user's primary calendar
      security:
        - OAuth2: []
      responses:
        '200':
          description: A list of calendar events
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        summary:
                          type: string
                        start:
                          type: object
                        end:
                          type: object
    post:
      operationId: createCalendarEvent
      summary: Create a new calendar event
      description: Creates a new event in the user's primary calendar
      security:
        - OAuth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary:
                  type: string
                description:
                  type: string
                start:
                  type: object
                end:
                  type: object
      responses:
        '200':
          description: The created calendar event
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  summary:
                    type: string

  /api/calendar/events/{eventId}:
    put:
      operationId: updateCalendarEvent
      summary: Update a calendar event
      description: Updates an existing event in the user's primary calendar
      security:
        - OAuth2: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary:
                  type: string
                description:
                  type: string
                start:
                  type: object
                end:
                  type: object
      responses:
        '200':
          description: The updated calendar event
    delete:
      operationId: deleteCalendarEvent
      summary: Delete a calendar event
      description: Deletes an event from the user's primary calendar
      security:
        - OAuth2: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmation of deletion

  /api/tasks/lists:
    get:
      operationId: getTaskLists
      summary: Get task lists
      description: Retrieves all task lists for the user
      security:
        - OAuth2: []
      responses:
        '200':
          description: A list of task lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        id:
                          type: string
    post:
      operationId: createTaskList
      summary: Create a new task list
      description: Creates a new task list
      security:
        - OAuth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '200':
          description: The created task list

  /api/tasks/lists/{taskListId}/tasks:
    get:
      operationId: getTasks
      summary: Get tasks in a list
      description: Retrieves all tasks in a specific task list
      security:
        - OAuth2: []
      parameters:
        - name: taskListId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of tasks
    post:
      operationId: createTask
      summary: Create a new task
      description: Creates a new task in a specific task list
      security:
        - OAuth2: []
      parameters:
        - name: taskListId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                notes:
                  type: string
                due:
                  type: string
      responses:
        '200':
          description: The created task

  /api/gmail/messages:
    get:
      operationId: getGmailMessages
      summary: Get recent emails
      description: Retrieves recent emails from the user's Gmail inbox with full message details
      security:
        - OAuth2: []
      responses:
        '200':
          description: A list of email messages with full details
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        threadId:
                          type: string
                        payload:
                          type: object
    post:
      operationId: sendEmail
      summary: Send an email
      description: Sends a new email message
      security:
        - OAuth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                subject:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Confirmation of sent email

  /api/gmail/messages/{messageId}:
    delete:
      operationId: deleteEmail
      summary: Move an email to trash
      description: Moves a specific email message to the trash
      security:
        - OAuth2: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmation of message moved to trash

  /api/drive/files:
    get:
      operationId: getDriveFiles
      summary: Get Drive files
      description: Retrieves a list of files from Google Drive with detailed information
      security:
        - OAuth2: []
      responses:
        '200':
          description: A list of files with details
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        mimeType:
                          type: string
                        webViewLink:
                          type: string
                        createdTime:
                          type: string
    post:
      operationId: createFile
      summary: Create a new file
      description: Creates a new file in Google Drive
      security:
        - OAuth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                mimeType:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: The created file

  /api/drive/files/{fileId}:
    get:
      operationId: getFile
      summary: Get file details
      description: Retrieves details of a specific file
      security:
        - OAuth2: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File details
    delete:
      operationId: deleteFile
      summary: Delete a file
      description: Deletes a specific file from Google Drive
      security:
        - OAuth2: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmation of file deletion

components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://ucandoai.netlify.app/auth/google
          tokenUrl: https://ucandoai.netlify.app/auth/google/callback
          scopes:
            'https://www.googleapis.com/auth/calendar': Read and write access to Calendar
            'https://www.googleapis.com/auth/tasks': Read and write access to Tasks
            'https://www.googleapis.com/auth/gmail.modify': Read and write access to Gmail
            'https://www.googleapis.com/auth/drive.file': Read and write access to Drive files